Work around a runc with readonlyPaths in the OCI spec
Generating an OCI spec including readonlyPaths and using the host
PID namespace, and then running it with any recent (RC91+) runc
as non-root will fail with an EPERM trying to remount the path as
read-only. We didn't hit this in Podman 2.x because we didn't
generate specs including readonlyPaths. Also, apparently none of
our runc testing upstream combines rootless containers and
pid=host, so this was not caught by CI.

Trivial workaround: don't generate specs including readonlyPaths
in the rootless + pid=host case. We can fix in runc upstream at a
later date, but we need something ready for RHEL quickly.

Signed-off-by: Matthew Heon <mheon@redhat.com>