Covert SpecGen values to be nullable where possible
SpecGen is our primary container creation abstraction, and is
used to connect our CLI to the Libpod container creation backend.
Because container creation has a million options (I exaggerate
only slightly), the struct is composed of several other structs,
many of which are quite large.

The core problem is that SpecGen is also an API type - it's used
in remote Podman. There, we have a client and a server, and we
want to respect the server's containers.conf. But how do we tell
what parts of SpecGen were set by the client explicitly, and what
parts were not? If we're not using nullable values, an explicit
empty string and a value never being set are identical - and we
can't tell if it's safe to grab a default from the server's
containers.conf.

Thus, this extremely large (and rather painful) refactoring. It's
likely only the first step (we have yet another level of
indirection in there, entities.ContainerCreateConfig, which I did
not touch yet, but will probably need a similar refactor to use
nullable values. That's not an API type, though, so we can do it
at any time, not just in a breaking change version like Podman 5.

[NO NEW TESTS NEEDED] This is just a refactor. Just a very, very
large refactor.

Signed-off-by: Matthew Heon <matthew.heon@pm.me>